+++ src/main/java/br/com/projects/main/repositories/ClienteRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.projects.main.entities.Cliente;
 
-public interface ClienteRepositorio extends JpaRepository<Cliente, Integer> {
+public interface ClienteRepositorio extends JpaRepository<Cliente, Long> {
 
 }
